state: move inline-help selectors to global folder
